# api/openapi.yaml
openapi: 3.0.3
info:
  title: BagiCore
  description: Go-based comic hosting catalog backend.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: 0.0.1
...
servers:
  - url: /api/v0
paths:
...
  /websites:
    get:
      tags:
        - Website
      summary: List website.
      operationId: listWebsite
      parameters:
        - name: page
          in: query
          description: Page number of results.
          schema:
            type: integer
        - name: limit
          in: query
          description: Maximum number of results.
          schema:
            type: integer
        - name: order_by
          in: query
          description: Sort results returned.
          schema:
            type: array
            items:
              type: string
      responses:
        '200':
          description: Website list.
          headers:
            X-Total-Count:
              schema:
                type: integer
              description: The total count of website with current filter.
            X-Pagination-Limit:
              schema:
                type: integer
              description: The last page number of website with current filter and limit.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Website'
        default:
          $ref: '#/components/responses/Default'
    post:
      tags:
        - Website
      summary: Add website.
      operationId: addWebsite
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NewWebsite'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/NewWebsite'
        required: true
      responses:
        '201':
          description: Website added.
          headers:
            Location:
              description: The path of new website.
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Website'
        default:
          $ref: '#/components/responses/Default'
      security:
        - BearerAuth: []
  /websites/{domain}:
    get:
      tags:
        - Website
      summary: Get website.
      operationId: getWebsite
      parameters:
        - name: domain
          in: path
          description: Domain name of website to return.
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Website gets.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Website'
        default:
          $ref: '#/components/responses/Default'
    patch:
      tags:
        - Website
      summary: Update website.
      operationId: updateWebsite
      parameters:
        - name: domain
          in: path
          description: Domain name of website to update.
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SetWebsite'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/SetWebsite'
        required: true
      responses:
        '200':
          description: Website updated.
          headers:
            Location:
              description: The path of updated website.
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Website'
        '204':
          description: Website unmodified.
        default:
          $ref: '#/components/responses/Default'
      security:
        - BearerAuth: []
    delete:
      tags:
        - Website
      summary: Delete website.
      operationId: deleteWebsite
      parameters:
        - name: domain
          in: path
          description: Domain name of website to delete.
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Website deleted.
        default:
          $ref: '#/components/responses/Default'
      security:
        - BearerAuth: []
  /websites/{domain}/tl-languages:
    post:
      tags:
        - Website
      summary: Add website TL language.
      operationId: addWebsiteTLLanguage
      parameters:
        - name: domain
          in: path
          description: Domain name of website.
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NewWebsiteTLLanguage'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/NewWebsiteTLLanguage'
        required: true
      responses:
        '201':
          description: Website TL language added.
          headers:
            Location:
              description: The path of new website TL language.
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WebsiteTLLanguage'
        default:
          $ref: '#/components/responses/Default'
      security:
        - BearerAuth: []
  /websites/{domain}/tl-languages/{ietf}:
    get:
      tags:
        - Website
      summary: Get website TL language.
      operationId: getWebsiteTLLanguage
      parameters:
        - name: domain
          in: path
          description: Domain name of website.
          required: true
          schema:
            type: string
        - name: ietf
          in: path
          description: IETF code of website TL language to return.
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Website TL language gets.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WebsiteTLLanguage'
        default:
          $ref: '#/components/responses/Default'
    patch:
      tags:
        - Website
      summary: Update website TL language.
      operationId: updateWebsiteTLLanguage
      parameters:
        - name: domain
          in: path
          description: Domain name of website.
          required: true
          schema:
            type: string
        - name: ietf
          in: path
          description: IETF code of website TL language to update.
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SetWebsiteTLLanguage'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/SetWebsiteTLLanguage'
        required: true
      responses:
        '200':
          description: Website TL language updated.
          headers:
            Location:
              description: The path of updated website TL language.
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WebsiteTLLanguage'
        '204':
          description: Website TL language unmodified.
        default:
          $ref: '#/components/responses/Default'
      security:
        - BearerAuth: []
    delete:
      tags:
        - Website
      summary: Delete website TL language.
      operationId: deleteWebsiteTLLanguage
      parameters:
        - name: domain
          in: path
          description: Domain name of website.
          required: true
          schema:
            type: string
        - name: ietf
          in: path
          description: IETF code of website TL language to delete.
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Website TL language deleted.
        default:
          $ref: '#/components/responses/Default'
      security:
        - BearerAuth: []
...
components:
  schemas:
    Object:
      type: object
      properties:
        id:
          type: integer
          format: int64
          x-go-type: uint
          x-go-name: ID
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
          nullable: true
      required:
        - id
        - createdAt
...
    Language:
      type: object
      allOf:
        - $ref: '#/components/schemas/Object'
        - type: object
          properties:
            ietf:
              type: string
              x-go-name: IETF
            name:
              type: string
          required: 
            - ietf
            - name
...
    Website:
      type: object
      allOf:
        - $ref: '#/components/schemas/Object'
        - type: object
          properties:
            domain:
              type: string
            name:
              type: string
            tlLanguages:
              type: array
              items:
                $ref: '#/components/schemas/Language'
              x-go-name: TLLanguages
            machineTL:
              type: boolean
              nullable: true
          required: 
            - domain
            - name
    NewWebsite:
      type: object
      properties:
        domain:
          type: string
        name:
          type: string
        machineTL:
          type: boolean
          nullable: true
      required: 
        - domain
        - name
    SetWebsite:
      type: object
      properties:
        domain:
          type: string
          nullable: true
        name:
          type: string
          nullable: true
        machineTL:
          type: boolean
          nullable: true
    WebsiteTLLanguage:
      type: object
      properties:
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
          nullable: true
        languageID:
          type: integer
          x-go-type: uint
        languageIETF:
          type: string
      required:
        - createdAt
        - languageID
        - languageIETF
    NewWebsiteTLLanguage:
      type: object
      properties:
        languageID:
          type: integer
          nullable: true
          x-go-type: uint
        languageIETF:
          type: string
          nullable: true
    SetWebsiteTLLanguage:
      type: object
      properties:
        languageID:
          type: integer
          nullable: true
          x-go-type: uint
        languageIETF:
          type: string
          nullable: true
...
    Error:
      type: object
      properties:
        error:
          type: object
          properties: 
            status:
              type: string
            message:
              type: string
          required:
            - status
            - message
      required:
        - error
  responses:
    Default:
      description: Unexpected error.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT